requirements.txt
gunicorn
djangorestframework
instagrapi
pillow